SW-4761 Create InventoryRouter and put all related views and sub components into the scenes directory. Move and rename the BatchBulkWithdrawView to the scenes directory and put the BatchWithdrawFlow into its own component that is shared in a couple routes
