--- src/components/Inventory/withdraw/index.tsx
@@ -1,4 +1,0 @@
-import BatchBulkWithdrawWrapperComponent from './BatchBulkWithdrawWrapperComponent';
-import SpeciesBulkWithdrawWrapperComponent from './SpeciesBulkWithdrawWrapperComponent';
-
-export { BatchBulkWithdrawWrapperComponent, SpeciesBulkWithdrawWrapperComponent };
