+++ src/scenes/InventoryRouter/EventDetailsModal.tsx
