+++ src/scenes/InventoryRouter/InventoryListByBatch.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { makeStyles } from '@mui/styles';
 import { TableColumnType } from '@terraware/web-components';
 import { CircularProgress, Container, Theme } from '@mui/material';
-import InventoryTable from 'src/components/InventoryV2/InventoryTable';
+import InventoryTable from 'src/scenes/InventoryRouter/InventoryTable';
 import EmptyStatePage from 'src/components/emptyStatePages/EmptyStatePage';
 import { SearchNodePayload, SearchResponseElement, SearchSortOrder } from 'src/types/Search';
 import { getRequestId, setRequestId } from 'src/utils/requestsId';
@@ -10,12 +10,12 @@ import { BE_SORTED_FIELDS, SearchInventoryParams } from 'src/services/NurseryInv
 import { useOrganization } from 'src/providers';
 import useDebounce from 'src/utils/useDebounce';
 import useForm from 'src/utils/useForm';
-import { InventoryFiltersUnion } from 'src/components/InventoryV2/InventoryFilter';
-import { BatchInventoryResult, InventoryResultWithBatchNumber } from 'src/components/InventoryV2';
+import { InventoryFiltersUnion } from 'src/scenes/InventoryRouter/InventoryFilter';
+import { BatchInventoryResult, InventoryResultWithBatchNumber } from 'src/scenes/InventoryRouter/InventoryV2View';
 import strings from 'src/strings';
 import { NurseryBatchService } from 'src/services';
 import Card from 'src/components/common/Card';
-import { isBatchEmpty } from 'src/components/InventoryV2/FilterUtils';
+import { isBatchEmpty } from 'src/scenes/InventoryRouter/FilterUtils';
 import isEnabled from 'src/features';
 
 const useStyles = makeStyles((theme: Theme) => ({
