+++ src/scenes/InventoryRouter/InventoryTable.tsx
@@ -6,12 +6,12 @@ import _, { isArray } from 'lodash';
 import strings from 'src/strings';
 import { useSessionFilters } from 'src/utils/filterHooks/useSessionFilters';
 import { SearchResponseElement, SearchSortOrder } from 'src/types/Search';
-import { InventoryFiltersUnion } from 'src/components/InventoryV2/InventoryFilter';
+import { InventoryFiltersUnion } from 'src/scenes/InventoryRouter/InventoryFilter';
 import { APP_PATHS } from 'src/constants';
 import Table from 'src/components/common/table';
 import { SortOrder } from 'src/components/common/table/sort';
-import { OriginPage } from 'src/components/InventoryV2/InventoryBatch';
-import Search from 'src/components/InventoryV2/Search';
+import { OriginPage } from 'src/scenes/InventoryRouter/InventoryBatchView';
+import Search from 'src/scenes/InventoryRouter/Search';
 import InventoryCellRenderer from './InventoryCellRenderer';
 import ProjectAssignTopBarButton from 'src/components/ProjectAssignTopBarButton';
 import isEnabled from 'src/features';
