+++ src/scenes/InventoryRouter/view/BatchesCellRenderer.tsx
