+++ src/scenes/InventoryRouter/view/DeleteBatchesModal.tsx
