+++ src/scenes/InventoryRouter/view/InventorySummaryForNursery.tsx
