+++ src/scenes/InventoryRouter/view/InventorySummaryForSpecies.tsx
