+++ src/scenes/InventoryRouter/view/QuantitiesMenu.tsx
